requirements.txt
camelot-py[cv]
pandas
openpyxl